+++ packages/app/src/elements/audit.ts
@@ -1,35 +1,17 @@
 import { translate as $l } from "@padloc/locale/src/translate";
-import { AuditResultType } from "@padloc/core/src/item";
-import { customElement, state, queryAll } from "lit/decorators.js";
+import { AuditResultType, VaultItem } from "@padloc/core/src/item";
+import { customElement } from "lit/decorators.js";
 import { css, html } from "lit";
-
 import { app } from "../globals";
 import { StateMixin } from "../mixins/state";
 import { Routing } from "../mixins/routing";
-import { AuditMixin } from "../mixins/audit";
-import { animateElement } from "../lib/animation";
 import { View } from "./view";
-import { ListItem } from "./items-list";
+import { iconForAudit, titleTextForAudit } from "../lib/audit";
+import { Vault } from "@padloc/core/src/vault";
 
 @customElement("pl-audit")
-export class Audit extends AuditMixin(StateMixin(Routing(View))) {
+export class Audit extends StateMixin(Routing(View)) {
     readonly routePattern = /^audit/;
-
-    @state()
-    private _reusedPasswords: ListItem[] = [];
-
-    @state()
-    private _weakPasswords: ListItem[] = [];
-
-    @state()
-    private _compromisedPasswords: ListItem[] = [];
-
-    @queryAll("pl-audit-list-item")
-    private _countElements: HTMLDivElement[];
-
-    handleRoute() {
-        this.audit();
-    }
 
     shouldUpdate() {
         return !!app.account;
@@ -54,8 +36,7 @@ export class Audit extends AuditMixin(StateMixin(Routing(View))) {
     ];
 
     render() {
-        const { _reusedPasswords, _weakPasswords, _compromisedPasswords } = this;
-
+        let items = app.auditedItems;
         return html`
             <div class="fullbleed vertical layout">
                 <header class="padded spacing center-aligning horizontal layout">
@@ -65,7 +46,7 @@ export class Audit extends AuditMixin(StateMixin(Routing(View))) {
                             this.dispatchEvent(new CustomEvent("toggle-menu", { composed: true, bubbles: true }))}
                     >
                         <div class="half-margined horizontal spacing center-aligning layout text-left-aligning">
-                            <pl-icon icon="shield-check"></pl-icon>
+                            <pl-icon icon="audit-pass"></pl-icon>
                             <div class="stretch ellipsis">${$l("Security Audit")}</div>
                         </div>
                     </pl-button>
@@ -74,15 +55,12 @@ export class Audit extends AuditMixin(StateMixin(Routing(View))) {
                     <div class="vertical spacing stretch">
                         <pl-scroller class="stretch">
                             <div class="counts">
-                                ${this._renderSection(_reusedPasswords, $l("Reused Passwords"), {
-                                    audit: AuditResultType.ReusedPassword,
-                                })}
-                                ${this._renderSection(_weakPasswords, $l("Weak Passwords"), {
-                                    audit: AuditResultType.WeakPassword,
-                                })}
-                                ${this._renderSection(_compromisedPasswords, $l("Compromised Passwords"), {
-                                    audit: AuditResultType.CompromisedPassword,
-                                })}
+                                ${Object.values(AuditResultType).map((type) =>
+                                    this._renderSection(
+                                        items.filter(({ item }) => item.auditResults.some((res) => res.type === type)),
+                                        type
+                                    )
+                                )}
                             </div>
                         </pl-scroller>
                     </div>
@@ -91,46 +69,45 @@ export class Audit extends AuditMixin(StateMixin(Routing(View))) {
         `;
     }
 
-    async audit() {
-        const { vaults } = this.state;
-
-        const { reusedPasswords, weakPasswords, compromisedPasswords } = await this.auditVaults(vaults);
-
-        // This makes the UI update
-        this._reusedPasswords = reusedPasswords;
-        this._weakPasswords = weakPasswords;
-        this._compromisedPasswords = compromisedPasswords;
-
-        this._countElements.forEach((countElement) => {
-            animateElement(countElement, { animation: "bounce" });
-        });
-    }
-
-    private _renderSection(listItems: ListItem[], label: string, routeParams: any) {
+    private _renderSection(listItems: { item: VaultItem; vault: Vault }[], type: AuditResultType) {
         return html`
-            <section class="box count" ?hidden=${listItems.length === 0}>
+            <section class="box count">
                 <h2 class="uppercase bg-dark border-bottom semibold center-aligning spacing horizontal layout">
-                    <div></div>
-                    <div>${label}</div>
-                    <div class="tiny bold tag">${listItems.length}</div>
+                    <pl-icon icon="${iconForAudit(type)}" class="left-margined"></pl-icon>
+                    <div>${titleTextForAudit(type)}</div>
+                    <div class="bold subtle">${listItems.length}</div>
                     <div class="stretch"></div>
+                    <pl-button class="subtle skinny transparent half-margined">
+                        <pl-icon icon="info-round"></pl-icon>
+                    </pl-button>
                 </h2>
-                <pl-list>
-                    ${listItems.slice(0, 5).map(
-                        (listItem) => html`
-                            <div class="list-item hover click" @click=${() => this.go(`items/${listItem.item.id}`)}>
-                                <pl-vault-item-list-item
-                                    .item=${listItem.item}
-                                    .vault=${listItem.vault}
-                                    class="stretch collapse"
-                                ></pl-vault-item-list-item>
-                            </div>
-                        `
-                    )}
-                </pl-list>
+                ${listItems.length
+                    ? html`
+                          <pl-list>
+                              ${listItems.slice(0, 5).map(
+                                  (listItem) => html`
+                                      <div
+                                          class="list-item hover click"
+                                          @click=${() => this.go(`items/${listItem.item.id}`)}
+                                      >
+                                          <pl-vault-item-list-item
+                                              .item=${listItem.item}
+                                              .vault=${listItem.vault}
+                                              class="none-interactive"
+                                          ></pl-vault-item-list-item>
+                                      </div>
+                                  `
+                              )}
+                          </pl-list>
+                      `
+                    : html`
+                          <div class="small double-padded subtle text-centering">
+                              <pl-icon icon="audit-pass" class="inline"></pl-icon> ${$l("Nothing Found")}
+                          </div>
+                      `}
                 <pl-button
                     class="slim margined transparent"
-                    @click=${() => this.go("items", routeParams)}
+                    @click=${() => this.go("items", { audit: type })}
                     ?hidden=${listItems.length < 6}
                 >
                     <div>${$l("Show All")}</div>
